third_party/blink/web_tests/jquery/core-expected.txt

ALERT: 667 tests of 667 passed, 0 failed.
